Cast bit shifts for socket flags to be uint32_t

+++ src/socket/source/socket_buffer.c
@@ -216,7 +216,7 @@ void socketBufferCb(SOCKET sock, uint8 u8Msg, void *pvMsg)
 				else {
 					/* Use flag to store spawn TCP descriptor. */
 					*(gastrSocketBuffer[sock].flag) &= ~SOCKET_BUFFER_FLAG_SPAWN_SOCKET_MSK;
-					*(gastrSocketBuffer[sock].flag) |= ((uint8)pstrAccept->sock) << SOCKET_BUFFER_FLAG_SPAWN_SOCKET_POS;
+					*(gastrSocketBuffer[sock].flag) |= (((uint32)pstrAccept->sock) << SOCKET_BUFFER_FLAG_SPAWN_SOCKET_POS);
 					*(gastrSocketBuffer[sock].flag) |= SOCKET_BUFFER_FLAG_SPAWN;
 				}
 			}
